Extract IncrementalMojoHelper and use it in the NpmMojo Incremental build now activated with the execution `incremental` configuration flag set to `true`
